recommendation_two_tower_full_cycle/services/training/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
psycopg2-binary==2.9.9
redis==5.0.1
numpy==1.26.2
pandas==2.1.3
torch==2.1.1
mlflow==2.9.1
boto3==1.33.6
prometheus-client==0.19.0
python-json-logger==2.0.7
scikit-learn==1.3.2
PyYAML==6.0.1
requests==2.31.0
